+++ examples/example1/main.py
@@ -1,0 +1,6 @@
+import quickstartlog as qslog
+qslog.debug('this is debug')
+qslog.info('this is info')
+qslog.warn('this is warning')
+qslog.error('this is error')
+qslog.info_ex(qslog.LIGHT_GREEN)('this is info with LIGHT_GREEN')